Move FormSetup type
