+++ src/services/ui/src/lib/common-types.ts
@@ -1,4 +1,5 @@
 import { Action } from "shared-types";
+import { ZodEffects, ZodObject } from "zod";
 
 export type PackageActionEndpoint = `/action/${Action}`;
 export type FormSubmissionEndpoint = "/submit";
@@ -12,4 +13,9 @@ export type AttachmentRecipe<S extends Record<string, unknown>> = {
   readonly name: keyof S["attachments"] | string;
   readonly label: string;
   readonly required: boolean;
+};
+
+export type FormSetup = {
+  schema: ZodObject<any> | ZodEffects<any>;
+  attachments: AttachmentRecipe<any>[];
 };
