+++ src/services/ui/src/pages/actions/IssueRai.tsx
@@ -4,7 +4,7 @@ import { zodResolver } from "@hookform/resolvers/zod";
 import { opensearch, PlanType } from "shared-types";
 import { ActionFormTemplate } from "@/pages/actions/template";
 import { useActionSubmitHandler } from "@/hooks/useActionFormController";
-import { FormSetup } from "@/pages/actions/setups";
+import { FormSetup } from "@/lib";
 
 const preSubmitMessage =
   "Once you submit this form, a confirmation email is sent to you and to the State.";
