--- src/services/ui/src/pages/actions/setups/_.ts
@@ -1,7 +1,0 @@
-import { ZodEffects, ZodObject } from "zod";
-import { AttachmentRecipe } from "@/lib";
-
-export type FormSetup = {
-  schema: ZodObject<any> | ZodEffects<any>;
-  attachments: AttachmentRecipe<any>[];
-};
