+++ src/services/ui/src/pages/actions/setups/index.ts
@@ -1,4 +1,3 @@
-export * from "./_"; // Types
 export * from "./setupIssueRai";
 export * from "./setupRespondToRai";
 export * from "./setupWithdrawRai";
